Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/4707



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@19890 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/transrt.cpp
@@ -1045,7 +1045,7 @@ static void writeStringInU8(FILE *out, const UnicodeString &s) {
         uint8_t  bufForOneChar[10];
         UBool    isError = FALSE;
         int32_t  destIdx = 0;
-        U8_APPEND(bufForOneChar, destIdx, sizeof(bufForOneChar), c, isError);
+        U8_APPEND(bufForOneChar, destIdx, (int32_t)sizeof(bufForOneChar), c, isError);
         fwrite(bufForOneChar, 1, destIdx, out);
     }
 }
